used api endpoints

+++ src/Components/RForm.js
@@ -28,46 +28,50 @@ function RForm({ btnName, setFunction }) {
   const handleSubmit = async (e) => {
     e.preventDefault();
     if (btnName === "Register") {
-      if (Object.entries(details).length > 0) {
-        if (details.password.length > 6) {
-          try {
-            setLoading(true);
-            await signup(details.email, details.password);
-            setResponse(true);
-            setMessage("You have registered successfully....");
-            setIsOpen(true);
-            setLoading(false);
-            const timedpush = () => {
-              setTimeout(() => history.push("/signIn"), 5000);
-            };
-            timedpush();
-            clearTimeout(timedpush);
-          } catch {
-            let choice = window.confirm("Would you like to sign in instead??");
-            if (choice) history.push("/signIn");
-            else setLoading(false);
-          }
+      if (Object.entries(details).length > 0 && details.password.length > 6) {
+        setLoading(true);
+        const res = await signup(details.email, details.password);
+        setLoading(false);
+        if (res.data === "success") {
+          setResponse(true);
+          setMessage("You have registered successfully");
+          setIsOpen(true);
+          const timedpush = () => {
+            setTimeout(() => history.push("/signIn"), 5000);
+          };
+          timedpush();
+          clearTimeout(timedpush);
+        } else {
+          let choice = window.confirm("Would you like to sign in instead??");
+          if (choice) history.push("/signIn");
+          else setLoading(false);
         }
+      } else {
+        setResponse(false);
+        setMessage("Please enter strong password....");
+        setIsOpen(true);
+        setLoading(false);
       }
     } else {
-      try {
-        setLoading(true);
-        const userCredentials = await login(details.email, details.password);
-        const email = userCredentials.user.email;
-        const id = userCredentials.user.uid;
-        const token = await userCredentials.user.getIdToken(true);
+      setLoading(true);
+      const res = await login(details.email, details.password);
+      const result = await res.data;
+      if (result.code === "auth/wrong-password") {
+        setResponse(false);
+        setMessage("Wrong email or password!!");
+        setIsOpen(true);
+        setDetails(info);
+        setLoading(false);
+      } else {
+        const email = result.user.email;
+        const id = result.user.uid;
+        const token = await result.user.stsTokenManager.refreshToken;
         setDetails(info);
         localStorage.setItem("authorized", token);
         history.push({
           pathname: `/passaver/${email.split("@", 1)}/`,
           state: { id },
         });
-      } catch {
-        setResponse(false);
-        setMessage("Wrong email or password!!");
-        setIsOpen(true);
-        setDetails(info);
-        setLoading(false);
       }
     }
   };
